キャッシュファイルがない場合で、Cache::deleteを呼び出すとWarningになるため修正 https://github.com/NetCommons3/NetCommons3/issues/1337

+++ Utility/NetCommonsCache.php
@@ -29,6 +29,13 @@ Cache::config('netcommons', $cacheSetting);
 class NetCommonsCache {
 
 /**
+ * キャッシュの設定値
+ *
+ * @var array
+ */
+	private $__setting = [];
+
+/**
  * キャッシュのデータ保持用
  *
  * @var array
@@ -65,9 +72,10 @@ class NetCommonsCache {
  * @return void
  */
 	public function __construct($cacheName, $isTest, $cacheType = 'netcommons') {
-		$this->__cacheName = $cacheName . '_' . NC3_VERSION;
+		$this->__cacheName = $cacheName . '_' . preg_replace('/\./', '_', NC3_VERSION);
 		$this->__isTest = $isTest;
 		$this->__cacheType = $cacheType;
+		$this->__setting = Cache::settings($this->__cacheType);
 
 		if (! $this->__isTest) {
 			$result = Cache::read($this->__cacheName, $this->__cacheType);
@@ -170,7 +178,11 @@ class NetCommonsCache {
 		$this->__data = [];
 
 		if (! $this->__isTest) {
-			$success = Cache::delete($this->__cacheName, $this->__cacheType);
+			if (file_exists($this->__setting['path'] . $this->__setting['prefix'] . $this->__cacheName)) {
+				$success = Cache::delete($this->__cacheName, $this->__cacheType);
+			} else {
+				$success = true;
+			}
 			if (! $success) {
 				$this->__triggerWarning('clear');
 			}
@@ -184,19 +196,18 @@ class NetCommonsCache {
  * @return void
  */
 	private function __triggerWarning($type) {
-		$cacheSetting = Cache::settings($this->__cacheType);
 		if ($type === 'delete') {
 			$message = __d(
 				'net_commons',
 				'Could not write to the %s cache file. Please check the file permissions.',
-				$cacheSetting['path'] . $cacheSetting['prefix'] . $this->__cacheName
+				$this->__setting['path'] . $this->__setting['prefix'] . $this->__cacheName
 			);
 		} else {
 			$message = __d(
 				'net_commons',
 				'The %s cache file could not be deleted. ' .
 					'Check the permissions of the file and delete the cache file.',
-				$cacheSetting['path'] . $cacheSetting['prefix'] . $this->__cacheName
+				$this->__setting['path'] . $this->__setting['prefix'] . $this->__cacheName
 			);
 		}
 		trigger_error($message, E_USER_WARNING);
